LOL-2 feat: initialize active game step

+++ data/src/main/java/com/yaini/data/enumerated/Spell.java
@@ -6,22 +6,22 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 public enum Spell {
-  BARRIER(21),
-  BOOST(1),
-  DOT(14),
-  EXHAUST(3),
-  FLASH(4),
-  HASTE(6),
-  HEAL(7),
-  MANA(13),
-  PORO_RECALL(30),
-  PORO_THROW(31),
-  SMITE(11),
-  SNOW_URF_SNOWBALL_MARK(39),
-  SNOWBALL(32),
-  TELEPORT(12),
-  ULT_BOOK_PLACE_HOLDER(54),
-  ULT_BOOK_SMITE_PLACE_HOLDER(55);
+  BARRIER(21L),
+  BOOST(1L),
+  DOT(14L),
+  EXHAUST(3L),
+  FLASH(4L),
+  HASTE(6L),
+  HEAL(7L),
+  MANA(13L),
+  PORO_RECALL(30L),
+  PORO_THROW(31L),
+  SMITE(11L),
+  SNOW_URF_SNOWBALL_MARK(39L),
+  SNOWBALL(32L),
+  TELEPORT(12L),
+  ULT_BOOK_PLACE_HOLDER(54L),
+  ULT_BOOK_SMITE_PLACE_HOLDER(55L);
 
-  private final Integer key;
+  private final Long key;
 }
